requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
python-dotenv
pydantic-settings
pydantic[email]
passlib[bcrypt]
python-multipart
python-jose[cryptography]
